add PeripheralCallbacks
