+++ Packages/com.teach310.core-bluetooth-for-unity/Runtime/NativeCallbacks/NativePeripheralCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20fa0ebfac1694eb187459a3fd262839
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
